Renamed product-handling in product-catalog-page to product-management to avoid confusion with product handing in the api folder
