+++ ecosphere-goods/src/components/pages/dashboard/product-catalog-page/ProductCatalogItem.js
@@ -1,7 +1,7 @@
 import React, { useContext } from 'react'
 import unitToDollarString from '../../../../utility-functions/unitToDollarString'
-import ConfirmDeleteDialog from './product-handling/ConfirmDeleteDialog';
-import ProductDialog from './product-handling/ProductDialog';
+import ConfirmDeleteDialog from './product-management/ConfirmDeleteDialog';
+import ProductDialog from './product-management/ProductDialog';
 import { FaEdit } from "react-icons/fa";
 import { ProductCatalogContext } from './ProductCatalogContext';
