+++ ecosphere-goods/src/components/pages/dashboard/product-catalog-page/product-management/ProductDialog.js
